transmission-scripts/requirements.txt
transmission-rpc
pyyaml